Add `pry: true` aliases for quick and simple debugging
Sets up aliases to examples/example groups in order to automatically
set a `binding.pry` statement in the state of the spec when an expectation
fails and debug why/what failed.

These failures usually occur deep inside the codebase and to debug it, you need
to concern yourself with adding the `binding.pry` statements **in the right places**
and ensuring you clean up afterwards.

With the `pry: true` metadata, these headaches/tediousness is gone

This is akin to using `fit` to set `focus: true` on an example.