Fix flaky test invite_user_modal/permission_lacking_current_project_spec.rb
Failing in https://github.com/opf/openproject/actions/runs/5420166041/jobs/9854007994

My hypothesis is that the flakiness occurs only because the two projects
have the same name, except the last character.

When the name of the second project is entered, before the last
character is entered, both projects are listed in the dropdown. Once the
last character is entered, there seems to be a race condition between
clicking on the option, and refreshing the list of options. The one
being clicked is the second project, but somehow the first project is
selected instead.

By waiting for the update of the options list, it does not occur
anymore.