Improving the README for dev env for Nix
Makes the headline clearer that it is about using Nix when developing
OpenProjec (and not using OpenProject in production).

As this is part of our public repo we should not invite readers to
directly contact developers. It makes sense the everyone in the
team knows who to contact for Nix support for development. However,
that is an internal invitation.