fix: align attribute groups across the application
Attribute groups on the storages and oauth application admin pages looked different than on the work package pages. This
was introduced by https://community.openproject.org/projects/openproject/work_packages/47680/activity

To align them, this commit reduces the scope of the `wp-attribute-group` css, instead applying those styles to the more
general `attributes-group` block. It also fixes the HTML structure of the backend-rendered attribute groups to align
with those in the frontend.

Ref https://community.openproject.org/work_packages/48640/activity