Fix another flaky test triggered by 3e18acec4b
Failing test was ./spec/features/work_packages/datepicker/datepicker_logic_spec.rb:264

The test was failing because clicking on the clear button of the
duration field did not work anymore. As the change detection occurs when
focusing out, when clicking on the clear button, the focus is lost and
the clear button is removed. This seems to cancel the click event on the
clear button which should have triggered setting the value to "".

Letting the clear button fisplayed when duration field is not focused,
but is selected (like for both start end and fields), fixes the test.

But duration field flickers between the durationDisplay and duration
texts (with and without " days" suffix). I tried removing the debounce
when the duration field is empty. It reduces the flickering
time but it still flickers.