Prevent rendering of delete buttons on global calendars page
As per requirements, the global index page for calendars, should not
render the delete buttons for each query on the table (in tandem with
other global index pages).

Also, extracted out some expectations into page objects in the
calendars#index page. This reads nicer to me and helps keep better
track of what is and isn't expected at a high level.