Ensure `folderModeValue` is updated to the selected radio button
The subscription in the `connect()` lifecycle hook uses `folderModeValue`
which is not updated until form submission when the template is re-rendered.

This was fine for initial rendering but for being able to update the `projectFolderSection`
dynamically when toggling between the different folder mode radio buttons, this
value needs to be updated to the currently selected folder mode in `updateForm()`.