Remove unnecessary `include` of `BoardsHelper`
I was mislead into this being needed during my testing as `BoardsHelper`
didn't seem to be available within the view but it seemed to be because
there was no `helpers` directory initially so it wasn't going to be autoloaded
until I stopped spring and restarted the rails server.

This is indeed unnecessary; Thanks Oliver.