Add `POST /projects/:project_id/boards` route
In order to be able to post from two different contexts and have
a failed validation on the form render back the appropriate form context, we
need create actions for both global and project boards. This follows the
same pattern as in other global and project pages.

**NOTE**: A path helper naming collision occured when attempting to add
the `create` action into the `resources` macro by using `only: %i[new create]`
as both the index and create action would end up with the same path helper, being
`project_work_package_boards_path`. Given we don't really use this path helper for
`POST`s throughout the application and build it up with `url_for`, I left a comment
explaining the reasoning for it and will proceed with using `url_for` as we do elsewhere.