Fix: `VersionBoardCreateService` should only create queries for direct project versions
After taking a look at the specs for the previously existing `version_board_spec` which
used the functionality from the Angular modals still, it only grabs versions that are directly
pertaining to the project in hand, regardless of the `sharing` attribute on the versions.

This commit updates the `VersionBoardCreateService` to match the current/previous behavior
of the Board Creation modal in order to stay in accordance with the expected behavior.

The way it was currently working with my implementation was that it would grab all shared versions
to that project (as returned by the versions api). The Angular Version Create Service appeared
to pluck out only those which were directly belonging to the project the board was being created
for, which I missed.