Replace special `POST /boards` route with `resources` macro
The special case for the `POST /boards` route was only needed due
to the collisions that would happen when sharing an index and show
route with the same path matcher. Once these routes were split, there
was no longer a path helper naming collission so we can ammend the
`resources` macro with `only: %i[create]` now as well, making better
use of it and cleaning up the routes.