Ensure Baseline is reset and deactivated when querying for `DEFAULT_TIMESTAMP`
The Baseline component was not watching for changes to the "pristine" state of
the current timestamp query being requested.

By watching for the pristine state being requested, now we can disable and reset
the Baseline details when selecting a builtin view (like "All open") which
doesn't include a timestamp (which would default the timestamp to 'PT0S').

However, when a saved query that includes a timestamp on it is selected due to
it have been saved that way, this will also cover the case of ensuring that
Baseline is not disabled and reset as the timestamp would not be the default.