Optimize `meetings_new_spec` by visiting `new` page directly
We don't really need to go via the index page every time. As long
as we can ensure that clicking on the "New meeting" button from it
routes properly in both the global and project-specific context, we
can just visit the new page directly for every other spec since
that's the real feature under test.