Remove db string length constraint on some fields
There is a mismatch between some database fields and model attributes
about the max length. This leads to a model being valid from a
ActiveRecord point of view, but failing to be inserted in database.

Such invalid constraints have been removed from the database, and the
remaining ones are controlled by a unit test.

This was also causing seeding issues for some languages as translated
names are longer than allowed.

Fixes https://community.openproject.org/wp/49583