Fix test failing in some time zones
The test finds the first time zone with compatible offset and uses it. But it does not use the canonical name which is used in code. For instance Europe/Amsterdam is not in the list returned by `time_zones_covering_1am_local_time` and it makes the test fail on my machine. Using the canonical name fixes it.