Correct Global Meetings Page sorting expectations
Prior to this commit, specs were passing but for the wrong reason. The
order in which the Project's are created in the test suite setup
happened to be the same order in which ids are assigned to each
Project record (and also the alphanumerical order given to each
Project's name).

By inverting this and naming the first created project alphanumerically
after the second project created, we can ensure the specs are either
passing or failing for the correct reasons.