Replace calls to `Rinku.auto_link` with the Rails wrapper method
As provided by the Rinku gem itself, it provides a helper that wraps
`Rinku.auto_link` with some goodies like auto-sanitizing if the
provided String is not marked as `html_safe` and re-marking as
`html_safe` on the result of `Rinku.auto_link` once the string is
acted upon by Rinku.

**NOTE**: The `AutolinkFilter` class has some special nuances in its
usage of css class setting and the `auto_link` helper doesn't really
fit as it would wrap all html_options in a `TagBuilder#tag_options`
and execute, which would end up wrapping the entire HTML and messes up
the Rinku rendering.