Refactor: Split `RoleSeeder` into `Project` and `Global` Role seeders
In order to more easily accomodate adding a seeder for
 `WorkPackageRole`s, splitting the catch-all `RoleSeeder` into a
seeder per type would help facilitate and simplify how we grok through
role seeding.

The translation file structure was also modified to more easily group
the different role types that are to be created whenever we start
adding more role types to be seeded.

The structure is now:
```yaml
<<role_type>>s:
  # role_data
```