Trigger download of chrome/chromedriver binaries in CI
When a new chrome or chromedriver version is available, the CI build
updates it. In fact, each of the 32 rspec processes will try to download
the new version and then run chromedriver. This leads to problems as the
chromedriver binary being run may still be downloaded by another rspec
process. This leads to the following error: "Text file busy".

This commit adds a step to the CI build that downloads the binaries
before the tests are run.