Create `DevelopmentData::WorkPackagesSeeder`
At the moment, we don't have any Work Packages in `[dev]` projects.

I would like to not pollute the Demo Data Work Packages with
memberships, as the Work Package memberships should be strictly
for development purposes.

Likewise, none of the development users are members in `demo-project`
or `scrum-project`, and this should also be the case for
Work Package memberships.