Fix submission of forms in CKEditor with turbo
The CKEditor form wrapper cancels and triggers resubmission of the form to be able to get all data from CKEditor
(which is asynchronous so the latest input from the user might not be present).

However turbo wraps submission events through clicks on submit and other events, so if we cancel the first submit,
the second submit will not go through turbo.

To ensure we can resubmit, we need to use turbo's own navigator, which again breaks on non-turbo pages.