Fix command to clear angular cache
The original command `find frontend/.angular/cache -type d -exec sh -c 'ls -dt
"$1"/*/ | tail -n +2 | xargs rm -r' sh {} \;` executes a command for each
directory found under `frontend/.angular/cache`. It fails in two cases:

* when the directory has no subdirectories, it fails with something like:
  ```
  ls: cannot access 'frontend/.angular/cache/16.2.2/babel-webpack/*/': No such file or directory
  ```

* when the directory has only one subdirectory, it's removed by `tail -n +2`
  and the rm command is called without any argument. The error looks like
  ```
  rm: missing operand
  Try 'rm --help' for more information
  ```

Another issue: `angular/cache/x.y.z/` contains 2 subdirectories: `babel-webpack`
and `angular-webpack`. The command could delete the oldest of both directories,
making the cache less effective. But this is mitigated by the lack of `-1`
argument for `ls`, so the directories may end up on the same line, in which case
both are kept, or not, in which case the oldest one is deleted.

The failures were silent when building the docker image as it's in the
`-exec` of `find`.

This commit replaces the command by a simpler one caring only about
keeping latest subdirectory of `angular-webpack`.