Fix race-condition in spec
There was a race-condition where the saving of the query was happening
as the update request for the added filter was also happening, causing
the frontend to not register the save as a successful creation but
as a successful update (from the added filter), causing the spec to
fail under conditions where it's execution was too fast.

The bcf_view_management spec is behaving a bit differently than others.
After adding the first filter, the filter count is still expected to
be "1" for some reason and starts incrementing with every other filter
added from then on. An injectable `initial_filter_count`.

**NOTE**: This is only a patch to aid in the flakiness that everyone
is experiencing but ultimately I would expect the "Save" button and
the "Save as" button to be disabled until every already performed
request is done (which would prevent this from happening to some users
just performing things at Capybara speed, or for some users under high
latency regions).