Create RSpec "service" spec generator
This generator is born from the tediousness of having to set up
boilerplate code every time I need to write a new service spec.

As a developer, I'd like to be able to leverage the beauty of
Rails generators to aid in my productivity.

When writing a new service spec, I'd like two things:

* Provide conventions for a good service spec: We have thorough
  shared examples for services that are in-line with the behavior
  provided by BaseServices. I'd like to be able to uncomment and keep
  the action (Create, Update, Delete) shared examples that apply for
  the service I'm testing and benefit from these already great
  specs (and not to mention... remember that they exist).

* Provide some good starting points for repetitive setup: Most service
  specs that deviate a bit from Base Services' shared examples
  require some custom setup. I'd like to get a head start at this
  and focus only on tuning and extending with very specific criteria
  that my spec requires.