Fix: `@component_wrapper_used` should be set on component and not target
The purpose of `@component_wrapper_used` in this scenario is to check
if the component that is to be inserted into the target has used
a `component_wrapper` in its template rendering.

This both:
* Fixes the issue where this will always raise an error as
  the target component is not being rendered (and doesn't need to be).

* Ensures that there is a unique ID set at the top-most level of
the turbo stream response and turbo can de-duplicate any existing
components that match the current ID of the component before
appending/prepending.