Add configurable `slowmo` option to cuprite specs
Have you ever wanted to debug a spec in headful mode but it's running
way too fast to tell what's going on? Adding a `SLOMO` environment
variable lets you add a delay (in seconds) to every action taken by
the driver.

Setting

```
SLOWMO=0.2
```

is a pretty good delay in order to be able to nicely inspect what's going
on and slows down the spec run enough to understand it.

Hope this helps everyone as it's helped me!