Enable Capybara screenshots S3 upload only with valid secrets
`CAPYBARA_AWS_ACCESS_KEY_ID` and `CAPYBARA_AWS_SECRET_ACCESS_KEY` are
stored as GitHub secrets and are only available to people with
collaborator access to the repository.

External collaborators and bots like dependabot will not be able to
upload screenshots to S3 on feature tests failure.

As the `docker-compose.ci.yml` forwards `CAPYBARA_AWS_ACCESS_KEY_ID` and
`CAPYBARA_AWS_SECRET_ACCESS_KEY`, they will be set to an empty string if
the value is not available due to access restrictions. That's why
checking for environment variable existence is not enough. It has to
check for presence (non-blank).