Fix: Register `WorkPackageMemberQuery`
The `MemberQuery` has a default_scope that checks a non-admin has a
certain set of permissions which are not required for the behavior of
the share modal (or visibility of work package shares for that matter).

Given that visibility is unscoped for the share modal, extracting a
separate query class that inherits from the MemberQuery sans the
default_scope made sense as a way to keep these two in sync minus
the varying visibility levels based on the current_user.