Feature 50849, reduce_by_writable_permissions shortcut for new models
For new models, per-field permissions are not considered, only the
default permission (e.g. add_work_package).
This is may not be a final solution, but at this moment I don't know how
to distinguish between per-field permissions for edit and create cases.

This fixes more tests, but copy WP scenario required also explicit
`status_id` in the merge block in CopyService#copied_attributes.
This was previously done for `parent_id` as well, so it's not exactly
a precedent.