Fix: Add proper wait expectation
Instead of sleeping for an arbitrary amount of time, adding an
expectation for the toast ensures that the default grid already
exists and can be queried for.

Without this, the spec is quite flaky as we can't guarantee that the
Grid exists within 0.5 seconds of the time the page loads.