Treat the condition dropdowns just as action selections are treated
Adding the action doesn't seem to be flaky as it's treated like an
ng-select autocomplete. When inspecting the failing screenshots on the
CI, it seems that the value on the project condition is not even set.

This tells me that there's probably a timing issue with ng-select that
a simple `fill_in :x, with: x` doesn't handle too well. However, I
haven't seen a single failure on the prior step which uses the
autocomplete logic for filling up the form field.

I'm giving this a try to see if just treating it like another ng-select
autocomplete stabilizes the spec.