Fix: Retry until ng-select is ready to accept events
It seems that  ng-select is not ready to receive events for some
reason. I'm unaware of a way to expect it to be ready to receive events
other attempting to set the value on the conditions multiple times.

Because the problem doesn't seem to be in finding the dropdown item to
click on in the autocompleter but in the expectation for the the value
to ACTUALLY be set, retrying on this sequence seems the way to go.

I did a bulk run while under 96% CPU usage with all passing.

This seems to be it! 🍾