Remove dead code
The code was not working properly: `property` is a string, it was
compared against a symbol, so `remaining_time` and `version` were always
considered writable.

It is a check in the contract which prevents the version from being
changed, and a check at the api level which prevents the remaining time
from being changed.

This code can be considered dead since it has not worked since more than
6 years (fd66a376e) and nobody noticed or complained.