Improve stability of custom_actions_spec
It was often failing with:

```
  1) Custom actions viewing workflow buttons
     Failure/Error:
       expect(page)
         .to have_css('.op-user-activity a.user-mention',
                      text: other_member_user.name,
                      wait: 10)

       expected to find css ".op-user-activity a.user-mention" within #<Capybara::Node::Element tag="div" path="//HTML[1]/BODY[1]/DIV[1]/DIV[1]/MAIN[1]/DIV[2]/OPENPROJECT-BASE[1]/DIV[1]/UI-VIEW[1]/WORK-PACKAGES-BASE[1]/DIV[1]/UI-VIEW[1]/WP-FULL-VIEW-ENTRY[1]/DIV[1]/DIV[2]/DIV[2]/DIV[1]/DIV[1]/OP-WP-TAB[1]/WP-ACTIVITY-TAB[1]/WORK-PACKAGE-COMMENT[1]/DIV[1]/DIV[1]/DIV[1]/DIV[1]"> but there were no matches
```

The activities were correctly fetched from the server, but the last one
was not displayed. It's because the activities tab displays acitivites
with this for loop: `*ngFor="let inf of visibleActivities;
trackBy:trackByIdentifier"`, and the identifier for `ActivityEntryInfo`
is composed of `href`, `version`, and `updatedAt` properties from the
activity.

The problem is that the `updatedAt` property only has second resolution,
and the two custom actions could be triggered within the same second. In
that case, the second one would not be displayed as the identifier did
not change.

The fix is to send milliseconds from the API for iso8601 date times.

This is only a partial fix as there are other flickering issues with the
custom_actions_spec at custom actions creation when setting the project
filter. Sometimes the selection does not work and it fails. Still, this
is a quality of dev improvement.