Make the Blankslate a BorderBox Row contextually
This allows maintaining the Border Box's header when no user matches
the applied filter. Prior to this, if you filtered for some Role and
there was no matching user that could be displayed, you'd lose the
header and no longer be able to filter until you closed and re-opened
the modal.