Fix flaky modules/costs/spec/features/timer_spec
This spec could fail with the following error:
```
Failure/Error: expect(page).to have_selector(".op-toast.-#{type}", text: message, wait: 5)
  expected to find css ".op-toast.-success" but there were no matches
```

Indeed when stopping the timer, the log time modal is displayed and the
test would click rather quickly on the save button, before the query to
gather available work packages actually triggers (because of the
debounce). This query is done to display available choices for the work
package input field dropdown.

When this happens, the PATCH `/api/v3/time_entries/:id` query completes
and sets the `ongoing` field to false. The query
`/api/v3/time_entries/:id/available_work_packages` is then triggered,
but will return a 404 because the time entry is not ongoing anymore and
thus is not visible to the user (in this test user only has
`:log_own_time` permission).

Because of this 404, an error message will be displayed in the toaster.
This will replace the toaster success message that is expected by the
test. That's why the test fails.

Waiting explicitly for the input field to display the work package
subject fixes the flakiness.