Pair work_package_role factories with what's actually seeded
This is just to maintain parity with the roles that are currently
default in OpenProject installations.

As a note, once we intend to make these work package roles
customizable, it's worth revisiting whether we want these specialized
factories with default permissions or not. As this is not going to
mimick what's actually being used in an instance.