Fix spec test.
spec/features/wiki/attachment_upload_spec.rb:107 got broken when using
`--headless=true`: when resizing the image and scrolling on the handle,
it scrolled out of bound to align the handle to the top of the page,
instead of aligning with the top of the text area.

By scrolling and aligning on the center vertically, it stays visible and
the drag handle can be dragged and dropped.