evaluate changes on journal creation only for supported associations
Similar to how it is already done for writing the journal changes later on, evaluating for any changes on associations (e.g. customizable_values)  is only done for associations that are
actually owned by the journaled object. For a work package, this excludes e.g. agenda_items, for meetings it will exclude e.g. customizable.