Use accessible selector for checking column header presence
It allows removing the `finicky` flag from the `add` method.

FWIW the `have_link` only works if the `a` tag has a `href` attribute,
and some column headers like "Overall costs" did not have it, which is
why another way of selecting them was needed. With the `:columnheader`
selector, it works fine and reads better.