Show only distinct users in members
The members page operates on the MemberQuery and members database table respectively.
With the sharing functionality, multiple memberships per users are now possible:

- 0 to 1 memberships on a project level (with at least one ProjectRole)
- 0 to N memberships on an entity/work package level (with at least one respective entity role).

This requires us to filter out which member roles we want to show to avoid duplicating member roles for each
of the above memberships.

The way we do this is by calling DISTINCT ON (user_id) on the members table with a condition of selecting records
with entity_id NULL first.

This way, if the user has a project membership, this record is retrieved first. Shared work packages are then shown in a separate
column using the `with_shared_work_packages_count` scope. If the user has no project memberships,
any first entity membership should be shown.