[42646] Fix untranslated "Spent time" widget name
The widget was registered with the identifier `time_entries_project`
which did not match the i18n key
`js.grid.widgets.time_entries_list.title`, so the presented title was
the `widget.options.name` value which is the translation at the widget
creation time. For instance if French was the locale when the widget is
added then the title would be in French.

Using the correct identifier `time_entries_list` fixes the issue and
translates it for all users.