[52384] Avoid updating ancestor twice
Processing ancestor twice leads to "Stale record" errors.

It happened in this scenario (see Bug #52384): when creating the child,
its attributes trigger a parent rescheduling. Once done, the parent is
part of the dependent results of the rescheduling, and as all results
(normal + dependent) are then processed by UpdateAncestorsService, the
parent gets processed twice: a first time when the child is processed,
and a second time when the parent is processed. As the parent has been
updated the first time, the second parent update will fail because this
parent instance is stale.

Solution is to store ids of already updated work packages, and skip the
update for the ones having already been updated.