Run feature specs with `parallel_split_test`
Not running unit specs with `parallel_split_test` as the penalty for
executing `before(:all)` hooks before every example is way too
noticeable in the overall execution time.

This however, is a negligible amount in the overall execution time for
feature specs. Being able to not pay a penalty for having a really long
feature spec file is a plus as this can help better balance the total
CI time.