Set locale in init-locale.ts and update it to avoid post formatting dates
Avoiding date postformatting is important, because in some languages the digit mapping is different.
Different digit maps cause issues while being sent to the server, we should always use the traditional Arabic numbers,
regardless of the chosen language.

Switching up the order of the loaded moment locales, so that english remains the default instead of german.