Fix CurrentUserService#loggedIn check
https://github.com/opf/openproject/pull/14672 introduced anonymous user output
to the current_user meta tag.

This results in the id being always present even if the user is not logged in.
Some of the application code depends on loggedIn checks by checking existence of this id.

These checks are now invalid, as anonymous is returned, but not logged in.

To fix this, a separate logged-in attribute is now used