Do not write or update sessions or their cookies for API access
Accessing the API currently does not, and should not update the user's session.
Still, currently we're using it to access the session for authenticating the user.

As a result, the session is loaded and we're currently outputting a Set-Cookie header as well as
writing the user session on every API request.

By using session_options[:skip], we can tell rack to avoid saving the session after the request