Add `md` sized left margin to ActionMenu
Since the scheme being used was "invisible", the button's padding was
mistaken for the margin between the author's name and the ActionMenu
show button. Upon hovering, this was noticeable and looked crowded.

This commit adds the default 8px margin between the two elements so
upon hovering the spacing is nice and tidy.