[30865] Show all work packages belonging to the version in status graph
In the graph widget, the default subprojectId filter is limiting the
work packages shown to the ones belonging to the version's project and
its subprojects. This is not always the desired behavior, especially
when the version is shared with other projects with 'hierarchy',
'tree', or 'system' sharing.

The fix opts for a projects filter including all visible projects for
the version when sharing is 'tree' or 'hierarchy' (which may have a
performance penalty when there are many projects), and a global scope
when sharing is 'system'.