[53584] Fix Portuguese languages mapping
https://community.openproject.org/wp/53584
https://community.openproject.org/wp/53586

When making Portuguese Brazilian and Portuguese Portugal translations
apart from each other in https://community.openproject.org/wp/53374, we
missed that the root key in the `pt-BR.yml` and `pt-PT.yml` files is
still `pt`. The backend is gracefully falling back to the `pt`
translations but the frontend just fails to translate and falls back to
the English strings, leading to mixed English and Portuguese in the user
interface.

A script fixes the root key in the `pt-BR.yml` and `pt-PT.yml` files.
This script is executed each time the translations are updated from
crowdin.