Statuses can no longer have a NULL default done ratio
It is 0 by default. This is to ensure that the total work, total
remaining work, and total % Complete are always consistent. Without it,
we could have Work being set and Remaining work being unset, which is
bad for the ancestors totals.

See https://community.openproject.org/wp/40749 for the detailed
specification.