Correctly format unsetting done_ratio in activity
Before, the `done_ratio` journal formatter was :decimal, which would
always convert nil to 0. And so unsetting the done_ratio would show as
"% Complete changed from 50 to 0".

Now it uses a dedicated :percentage formatter, which will render nil as
expected and also add a "%" sign at the end of the value. And so
unsetting the done_ratio now shows as "% Complete deleted (30%)".