Fix table helpers: do not ignore header if first row value is empty
When using `match_table` helper, some column were missing on the output.
It's because the headers where read from the first row, and the ones
without any value were stripped out. As this information is used to know
how to pull column data from work packages, the data was missing.