Ensure modals can be previewed in the context of create forms
Given the requests for previews are going to the `edit` action, this
relies on a path that includes a `:work_package_id` in the request's
path. Since the work package in a create context is unpersisted, there
isn't an id to be provided here.

This commit mimicks the behavior in the Angular side of the application
by setting the `:work_package_id` to "new" in create contexts.