[#53344] Navigate to the default query page when other users query has been made private
https://community.openproject.org/work_packages/53344

This covers an edge case, when an adminitrator makes another user's query private.
In this case, the query becomes accessible only by its creator. We will try to reload
the query after the save happens, but since the query is not accessible, a confusing error
message is rendered. In order to avoid this error message, we are redirecting the admin to
the default query page if the query becomes private and not accessible.