avoid instantiating ProjectRepresenter if it is not embedded anyway
In case of a collection, ebed_links will be false. But although it will prevent the embedded elements from being rendered, it does not stop the getter lambda from being executed resulting in ProjectRepresenters being created and then instantiated. Since the created class depends on the custom fields being active for the project, the custom fields are determined which can be costly if done once for each project.