Update the aac patch to disable custom fields on the before_commit hook and also disable non available custom fields.
We have a few callbacks that interfered with the
disable_custom_fields_with_empty_values being on the after_create
callback. For example the CustomValue model also has an after_create
callback activate_custom_field_in_customized_project which overwrites
the changes we do in the disable_custom_fields_with_empty_values. This
makes all the custom fields with a custom value to be activated,
regardless of the changes we do in the
disable_custom_fields_with_empty_values. Since we want the
disable_custom_fields_with_empty_values to have the final word, it is
moved into a before_commit hook, which will be executed after the
activate_custom_field_in_customized_project.