Fix share/filter_spec
For some reason this stopped working on the live-preview branch. Upon
inspecting the behavior of the method with some debuggers, it seemed
correct that the specs were to fail as it does fail as well upon
manual testing. What's interesting is that it only fails:

1. On subsequent modal updates
2. Not on initial modal loading

This is the best fix I could come up with in the meantime. I'm not
aware of a reason within the current branch as to why this stopped
working.