Attempt to fix flaky specs where the toast might disappear before the page finished loading
Let's consider the folliwng scenario: We are saving a form by clicking
the submit button. The page redirects, the toast message is shown, but
the page still loads some requests in the background. Capybara will not
consider the page submission with the "click_on" command done, unless
the page is loaded. Hence it will not go to the next expectation unless
the page is loaded. If the page takes too long to load, the toast
message might already disappear by the time capybara runs the next
expectation for the toast's presence. This will make the expectation
fail.