[54500] Refresh filters cache before building the query in job
https://community.openproject.org/wp/54500

If not refreshed, the only known and usable custom fields for filters
and groupbys are the ones that were present when the server was started
and a query was first built, because some classes are created on the fly
for the custom fields. Queries using custom fields added after will fail
to be found.

Even worse, in a multi-tenant environment, a cached custom field class
could be found and returned, but could have been generated from the
custom fields of another tenant, because the cache is global.

Calling `CostQuery::Cache.check` will check if the cache is outdated by
checking the number of custom fields and the last updated timestamp of
the custom fields. If the cache is outdated, the custom field `GroupBy`
and `Filter` classes will be deleted and recreated.